refactor: move file resolver into shared.
