+++ _shared/__init__.py
@@ -1,0 +1,3 @@
+# Copyright (c) 2023 Benjamin Mummery
+
+"""Common resources that are used by multiple hooks."""
