+++ tests/shared/unit/__init__.py
@@ -1,0 +1,4 @@
+# Copyright (c) 2023 Benjamin Mummery
+
+"""Required for sharing utilities across test files.
+"""
